Revert "Show wet track icon on track times (#39)"
This reverts commit 3bde47b82c65e7c0f8c1e0ac96082649bf5b6f29.
