--- copy-static/site.webmanifest
@@ -1,19 +1,0 @@
-{
-  "name": "IPESC",
-  "short_name": "IPESC Data Hub",
-  "icons": [
-    {
-      "src": "public/android-chrome-192x192.png",
-      "sizes": "192x192",
-      "type": "image/png"
-    },
-    {
-      "src": "public/android-chrome-512x512.png",
-      "sizes": "512x512",
-      "type": "image/png"
-    }
-  ],
-  "theme_color": "#ffffff",
-  "background_color": "#ffffff",
-  "display": "standalone"
-}
