+++ src/components/changelog/index.tsx
@@ -37,18 +37,6 @@ export const Changelog: FC = () => {
 const ChangeList: FC = () => {
   return (
     <>
-      <Version v='0.8.0'>
-        <Change type='minor'>
-          <LinkTo page='calculator'>Fuel calculator</LinkTo> accepts stint timer on
-          endurance mode.
-        </Change>
-        <Change type='minor'>Added favicon icon ー at last :)</Change>
-        <Change type='minor'>
-          Changed the way <LinkTo page='drivers'>driver rank</LinkTo> is calculated to
-          base event times for 100% to be the best of the event, not the track (as weather
-          conditions might affect it)
-        </Change>
-      </Version>
       <Version v='0.7.4'>
         <Change type='data'>Event data updated with the full S8 data.</Change>
       </Version>
