+++ src/components/fuel-calculator/stint-simulator/index.tsx
@@ -1,10 +1,11 @@
-import { FC } from 'react';
+import { ChangeEventHandler, FC, ReactNode } from 'react';
 import {
   Box,
   Flex,
   HStack,
   Heading,
   ListItem,
+  Select,
   Text,
   Tooltip,
   UnorderedList,
@@ -12,9 +13,10 @@ import {
 import { InfoIcon } from '@chakra-ui/icons';
 
 import { NumericInput } from 'components/numeric-input';
+import { msToTime } from 'utils/time';
 
-import { useStintSimulator } from './hooks';
-import { StintSummary } from './stint-summary';
+import { Summary, Props as SummaryProps } from '../summary';
+import { useStintSimulator, LapData, SimulationData } from './hooks';
 
 export interface Props {
   raceDuration: number | undefined;
@@ -30,8 +32,6 @@ export const StintSimulator: FC<Props> = (props) => {
     simulations,
     pitWindowHours,
     pitWindowMins,
-    stintDurationHours,
-    stintDurationMins,
     minPitstops,
     maxPitstops,
     pitstopSecs,
@@ -42,8 +42,6 @@ export const StintSimulator: FC<Props> = (props) => {
     updateMinPitstops,
     updateMaxPitstops,
     updatePitstopSecs,
-    updateStintDurationHours,
-    updateStintDurationMins,
     updateLapDegradationSecs,
     updateSelectedSimulation,
   } = useStintSimulator(props);
@@ -61,12 +59,6 @@ export const StintSimulator: FC<Props> = (props) => {
         mins={pitWindowMins}
         updateHours={updatePitWindowHours}
         updateMins={updatePitWindowMins}
-      />
-      <StintDuration
-        hours={stintDurationHours}
-        mins={stintDurationMins}
-        updateHours={updateStintDurationHours}
-        updateMins={updateStintDurationMins}
       />
       <PitstopTime value={pitstopSecs} update={updatePitstopSecs} />
       <LapDegradation value={lapDegradationSecs} update={updateLapDegradationSecs} />
@@ -105,40 +97,6 @@ const PitWindow: FC<{
           min={0}
           label='Mins'
           onChange={updateMins}
-        />
-      </Flex>
-    </Box>
-  );
-};
-
-const StintDuration: FC<{
-  hours: string | undefined;
-  mins: string | undefined;
-  updateHours: (value: string) => void;
-  updateMins: (value: string) => void;
-}> = ({ hours, mins, updateHours, updateMins }) => {
-  return (
-    <Box marginTop={4}>
-      <Heading marginBottom={1} size='xs'>
-        Stint Duration
-      </Heading>
-      <Flex justifyContent='space-between'>
-        <NumericInput
-          placeholder='00'
-          value={hours}
-          min={0}
-          onChange={updateHours}
-          marginRight={2}
-          label='Hrs'
-          width='50%'
-        />
-        <NumericInput
-          placeholder='00'
-          value={mins}
-          min={0}
-          onChange={updateMins}
-          label='Mins'
-          width='50%'
         />
       </Flex>
     </Box>
@@ -280,5 +238,247 @@ const LapDegradation: FC<{
         onChange={update}
       />
     </Box>
+  );
+};
+
+const Simulation: FC<{
+  simulation: SimulationData | undefined;
+}> = ({ simulation }) => {
+  const lapBars: ReactNode[] = [];
+  if (simulation) {
+    const { laps } = simulation;
+    for (let i = 0; i < laps.length; i++) {
+      lapBars.push(
+        <Lap key={i} index={i} fuelPerLap={simulation.fuelPerLap} {...laps[i]} />
+      );
+    }
+  } else {
+    lapBars.push(
+      <Flex key='no-laps' alignItems='center'>
+        <Text align='center' color='gray.400'>
+          Enter{' '}
+          <Text as='span' fontWeight='medium'>
+            Race
+          </Text>{' '}
+          and{' '}
+          <Text as='span' fontWeight='medium'>
+            Lap
+          </Text>{' '}
+          times to visualize data and make sure the number of{' '}
+          <Text as='span' fontWeight='medium'>
+            Pit Stops
+          </Text>{' '}
+          and the{' '}
+          <Text as='span' fontWeight='medium'>
+            Fuel Tank Capacity
+          </Text>{' '}
+          have reasonable values.
+        </Text>
+      </Flex>
+    );
+  }
+
+  return (
+    <Box marginTop={4}>
+      <Flex
+        height='102px'
+        width='100%'
+        border='1px solid'
+        borderColor='gray.200'
+        position='relative'
+      >
+        {lapBars}
+      </Flex>
+    </Box>
+  );
+};
+
+const Lap: FC<
+  LapData & {
+    index: number;
+    fuelPerLap: number | undefined;
+  }
+> = ({ index, ...lapData }) => {
+  const { width, height, pits, windowOpen } = lapData;
+  const tooltip = <LapDetails index={index} {...lapData} />;
+
+  return (
+    <Tooltip hasArrow label={tooltip}>
+      <Flex
+        data-group
+        height='100%'
+        flexGrow={width}
+        alignItems='flex-end'
+        backgroundColor={pits ? 'blue.50' : windowOpen ? 'green.50' : 'white'}
+        borderLeft={index === 0 ? undefined : '1px'}
+        borderColor={'white'}
+        _hover={{
+          backgroundColor: pits ? 'blue.100' : windowOpen ? 'green.100' : 'gray.50',
+        }}
+      >
+        <Box
+          height={`${100 * height}%`}
+          flexGrow={1}
+          backgroundColor={pits ? 'blue.200' : 'orange.300'}
+          _groupHover={{ backgroundColor: pits ? 'blue.400' : 'orange.400' }}
+        />
+      </Flex>
+    </Tooltip>
+  );
+};
+
+const LapDetails: FC<LapData & { index: number; fuelPerLap?: number }> = ({
+  index,
+  fuelPerLap,
+  lapTime,
+  delta,
+  raceTime,
+  fuel,
+  stint,
+  stintLap,
+  pits,
+}) => {
+  const labelColor = 'whiteAlpha.800';
+  const lapTitle = (
+    <Text>
+      <Text as='span' color={pits ? 'blue.400' : 'orange.400'} fontWeight='bold'>
+        Lap {index + 1}
+      </Text>
+      <Text as='span' color='whiteAlpha.700' ml={2}>
+        (S{stint + 1} L{stintLap + 1})
+      </Text>
+    </Text>
+  );
+  const deltaInfo = delta ? (
+    <Text as='span' size='sm'>
+      <Text as='span' color='red.400' ml={2}>
+        <Text as='span' size='xs'>
+          ▲
+        </Text>{' '}
+        {msToTime(delta, { min: false })}
+      </Text>{' '}
+      s
+    </Text>
+  ) : undefined;
+  const currentLapTime = (
+    <Text>
+      <Text as='span' color={labelColor}>
+        LapTime:
+      </Text>{' '}
+      {msToTime(lapTime)}
+      {deltaInfo}
+    </Text>
+  );
+  const ellapsedRaceTime = (
+    <Text>
+      <Text as='span' color={labelColor}>
+        Racetime:
+      </Text>{' '}
+      {msToTime(raceTime, { ms: false })}
+    </Text>
+  );
+  const fuelInfo = fuel && (
+    <Text as='span'>
+      <Text as='span' color={labelColor}>
+        Remaining fuel:
+      </Text>{' '}
+      <Text as='span' color={fuel < (fuelPerLap || 1) ? 'red' : undefined}>
+        {fuel.toFixed(2)}
+      </Text>{' '}
+      L{' '}
+      {fuelPerLap && (
+        <Text as='span' color='whiteAlpha.700'>
+          ({Math.floor(fuel / fuelPerLap)} laps)
+        </Text>
+      )}
+    </Text>
+  );
+  const pittingFuel = typeof pits === 'number' && (
+    <Text as='span' color='green.400' ml={2}>
+      +{Math.round(pits)} L
+    </Text>
+  );
+  const pitting = pits && <Text>Box{pittingFuel}</Text>;
+
+  return (
+    <Box marginTop={2}>
+      {lapTitle}
+      {currentLapTime}
+      {pitting}
+      {fuelInfo}
+      {ellapsedRaceTime}
+    </Box>
+  );
+};
+
+const StintSummary: FC<{
+  simulations: SimulationData[];
+  selectedIndex: number;
+  update: ChangeEventHandler<HTMLSelectElement>;
+}> = ({ simulations, selectedIndex, update }) => {
+  const simulation = simulations[selectedIndex];
+
+  const stintDescription: SummaryProps['items'] = [
+    simulation && {
+      label: 'Race start',
+      data: simulation?.initialFuel
+        ? `${simulation?.initialFuel.toFixed(0)} L`
+        : undefined,
+    },
+    ...(simulation?.stops || []).map(({ lap, fuel }, i) => ({
+      label: `Pit stop #${i} at lap ${lap}`,
+      data: fuel ? `${Math.round(fuel)} L` : undefined,
+    })),
+  ];
+
+  return (
+    <Summary
+      items={[
+        {
+          label: 'Total Fuel Needed',
+          data: simulation?.totalFuel
+            ? `${simulation?.totalFuel.toFixed(0)} L`
+            : undefined,
+          main: true,
+        },
+        {
+          label: 'Total Laps',
+          data: simulation?.laps.length,
+        },
+        {
+          label: 'Expected Race Time',
+          data: msToTime(simulation?.raceTime, { ms: false }),
+        },
+        ...stintDescription,
+      ]}
+    >
+      <SimulationsSelector
+        simulations={simulations}
+        update={update}
+        selectedIndex={selectedIndex}
+      />
+      <Simulation simulation={simulation} />
+    </Summary>
+  );
+};
+
+const SimulationsSelector: FC<{
+  simulations: SimulationData[];
+  selectedIndex: number;
+  update: ChangeEventHandler<HTMLSelectElement>;
+}> = ({ simulations, selectedIndex, update }) => {
+  if (!simulations.length) return;
+
+  const value = simulations[selectedIndex]?.stops.length;
+
+  return (
+    <Select value={value} onChange={update}>
+      {simulations.map((s) => (
+        <option key={`${s.stops.length}-${s.raceTime}`} value={s.stops.length}>
+          {s.stops.length} {s.stops.length === 1 ? 'stop' : 'stops'} 【{s.laps.length}{' '}
+          laps | {msToTime(s.raceTime)}】
+        </option>
+      ))}
+    </Select>
   );
 };
