+++ src/components/fuel-calculator/stint-simulator/simulate.ts
@@ -6,7 +6,6 @@ export interface SimulationParameters {
   windowDuration?: number;
   lapDegradationSecs?: number;
   pitstopSecs?: number;
-  stintDurationSecs?: number;
   fuelPerLap?: number;
   fuelTank?: number;
   extraLaps?: number;
@@ -22,7 +21,6 @@ export function simulate(
     pitstopSecs,
     fuelPerLap,
     fuelTank,
-    stintDurationSecs,
     extraLaps,
   }: SimulationParameters
 ): SimulationData | undefined {
@@ -52,9 +50,8 @@ export function simulate(
 
   const deltaPerLap = (1000 * (lapDegradationSecs || 0)) / 10;
   let raceTime = 0;
-  let lap = 1;
+  let lap = 0;
   let stintLap = 0;
-  let stintTime = 0;
   while (raceTime < raceDuration) {
     const lapDelta = deltaPerLap * stintLap;
     let currentLapTime = lapTime + lapDelta;
@@ -78,19 +75,11 @@ export function simulate(
       pits,
     });
 
-    stintLap++;
-    stintTime += currentLapTime;
-
     if (pits) {
       stops.push({ lap });
-
-      // if the stint is more than the permitted, this simulation is not valid
-      if (stintDurationSecs && stintTime > stintDurationSecs) {
-        return;
-      }
-
       stintLap = 0;
-      stintTime = 0;
+    } else {
+      stintLap++;
     }
 
     lap++;
@@ -127,10 +116,9 @@ export function simulate(
     let stop = 0;
     for (let i = 0; i < laps.length; i++) {
       const lap = laps[i];
-      lap.fuelOnStart = fuel;
+      lap.fuel = fuel;
       lap.height = fuel / fullTank;
       fuel -= fuelPerLap;
-      lap.fuelOnEnd = fuel;
 
       if (lap.pits) {
         const nextStintNeeded = stintFuel[lap.stint + 1] + extraFuel;
