+++ src/entries/index/index.html
@@ -10,11 +10,7 @@
           name="viewport"
           content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no"
         />
-        <link rel="apple-touch-icon" sizes="180x180" href="public/apple-touch-icon.png" />
-        <link rel="icon" type="image/png" sizes="32x32" href="public/favicon-32x32.png" />
-        <link rel="icon" type="image/png" sizes="16x16" href="public/favicon-16x16.png" />
-        <link rel="manifest" href="site.webmanifest" />
-        <title>IPESC Data Hub</title>
+        <title>IPESC data hub</title>
       </head>
       <body>
         <div id="root"></div>
