+++ src/utils/format-data.ts
@@ -1,0 +1,3 @@
+export function formatRatioAsPctg(n: number, decimalPoints = 2): string {
+  return (n * 100).toFixed(decimalPoints) + '%';
+}
